requirements.txt
crewai==0.28.8
langchain==0.1.20
langchain-google-genai==0.0.11
google-genai
streamlit
python-dotenv
requests
pandas
google-generativeai
reportlab
tenacity